refactor: 프로젝트 파일 구조 변경
