+++ src/components/common/Container.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import type React from "react";
-import Navigation from "@/components/Navigation";
+import Navigation from "@/components/common/Navigation";
 
 type ContainerProps = {
   children: React.ReactNode;
