+++ src/pages/Gift/Page.tsx
@@ -1,9 +1,9 @@
 import Container from "@/components/common/Container";
-import Category from "@/components/Category";
-import Ranking from "@/components/Ranking";
-import Friend from "@/components/Friend";
+import Category from "@/pages/Gift/components/Category";
+import Ranking from "@/pages/Gift/components/Ranking";
+import Friend from "@/pages/Gift/components/Friend";
 import Divider from "@/components/common/Divider";
-import Banner from "@/components/Banner";
+import Banner from "@/pages/Gift/components/Banner";
 
 const Gift = () => {
   return (
