+++ src/pages/Gift/components/Banner.tsx
