+++ src/pages/Gift/components/Friend.tsx
